requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==4.1.2
djangorestframework==3.14.0
idna==3.4
lxml==4.9.1
mypy-extensions==0.4.3
numpy==1.23.4
pandas==1.5.1
pathspec==0.10.1
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
urllib3==1.26.12
